Install react (#9)
Basic bare bones of the React infrastructure being added to the App,
renders the <App> component in the room page.

The polling function was not being shut down properly because React
isn't handling the routing, the componentWillUnmount method wasn't being
called. That would be okay except that Turbolinks, in not truly navigating
to another page, would not clear out the interval either.

Removing turbolinks b/c of unexpected behaviors. As per the usual.